+++ Front/front-react-project/src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Home from "./Home.tsx";
-import InputDreamPage from "./InputDreamPage.tsx";
+import InputDreamPage from "./page/InputDreamPage.tsx";
 
 function App() {
   return (
