+++ Front/front-react-project/src/input-form/InputSakuraCardItem.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Box, TextField } from "@mui/material";
-import "./style.css";
+import "../style.css";
 
 const InputSakuraCardItem = () => {
   return (
