+++ Front/front-react-project/src/input-form/OtherUserSakuraCard.tsx
@@ -1,6 +1,6 @@
 import { Box, Typography } from "@mui/material";
 import React from "react";
-import "./style.css";
+import "../style.css";
 
 type OtherUserPostProps = {
   dream: string;
