+++ Front/front-react-project/src/input-form/SakuraCardArea.tsx
